% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCaret-package.R
\docType{package}
\name{autoCaret}
\alias{autoCaret}
\alias{autoCaret-package}
\title{autoCaret: Create one-click binary classification models}
\description{
Functions and a graphical interface to 'automagically' create high performing binary classification models
}

