% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCaret.R
\name{autoTrainTestSplit}
\alias{autoTrainTestSplit}
\title{Function to wrap and automate datasplitting into train and test sets}
\usage{
autoTrainTestSplit(df, y, set_seed = 1234, side_effects = FALSE, p = 0.8,
  ...)
}
\arguments{
\item{df}{The dataframe to split into training and test}

\item{y}{The response variable in thataset}

\item{set_seed}{The seed to use to ensure that examples are reproducible}

\item{side_effects}{Whether or not to write training and test dataframe to the global env.}

\item{p}{The percentage of data that goes to training}
}
\description{
Function to wrap and automate datasplitting into train and test sets
}
\examples{

train_test_split(stroke.death.predicition, Y)

}

