% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCaret.R
\name{is.autoCaret}
\alias{is.autoCaret}
\title{Check if an object is a autoCaret object}
\usage{
is.autoCaret(object)
}
\arguments{
\item{object}{an R object}
}
\description{
Check if an object is a autoCaret object
}

