% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCaretUI.R
\docType{data}
\name{model_descriptions}
\alias{model_descriptions}
\title{autoCaret Add-in}
\format{An object of class \code{data.frame} with 6 rows and 6 columns.}
\usage{
model_descriptions
}
\value{
The function launches a shiny app in the system web browser.
}
\description{
This function launches a shiny app in a web browser in order to
run the autoCaret UI
}
\keyword{datasets}

