% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCaret.R
\name{predict.autoCaret}
\alias{predict.autoCaret}
\title{Extends predict functionality for autoCaret objects}
\usage{
predict.autoCaret(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a \code{\link{autoCaret}}}

\item{...}{optional additional parameters. Should correspond to caret::predict.train(object, newdata = NULL, type = "raw", na.action = na.omit)}
}
\description{
Extends caret predict.train to \code{\link{autoCaret}}
}
\examples{
\dontrun{
}

}

