% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCaret.R
\name{summary.autoCaret}
\alias{summary.autoCaret}
\title{Summarize the results of an autoCaret object for the user}
\usage{
\method{summary}{autoCaret}(object, ...)
}
\arguments{
\item{object}{a \code{\link{autoCaret}}}

\item{...}{optional additional parameters.}
}
\description{
Summarize an autoCaret object
}
\examples{
\dontrun{
}

}

