% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCaretUI.R
\name{autoCaretUI}
\alias{autoCaretUI}
\title{autoCaret Add-in}
\usage{
autoCaretUI(obj = NULL, var_name = NULL)
}
\value{
The function launches a shiny app in the system web browser.
}
\description{
This function launches a shiny app in a web browser in order to
run the autoCaret UI
}

